Move acceptance and tools to root
This sets up the acceptance tests as a separate go module and moves
acceptance and tools from `internal` to repository root to better
reflect them as modules.
