+++ .github/workflows/codeql.yaml
@@ -56,7 +56,8 @@ jobs:
       - name: Download go dependencies
         run: |
           go mod download
-          go mod download -modfile internal/tools/go.mod
+          go mod download -modfile tools/go.mod
+          go mod download -modfile acceptance/go.mod
 
       - name: Initialize CodeQL
         uses: github/codeql-action/init@v2
