+++ .github/workflows/lint.yaml
@@ -46,7 +46,10 @@ jobs:
           cache-dependency-path: '**/go.sum'
 
       - name: Download go dependencies
-        run: go mod download
+        run: |
+          go mod download
+          go mod download -modfile tools/go.mod
+          go mod download -modfile acceptance/go.mod
 
       - name: Lint
         run: make lint
