+++ acceptance/cli/cli_test.go
@@ -14,8 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build unit
-
 package cli
 
 import (
