+++ acceptance/tekton/bundles.go
@@ -31,7 +31,7 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/stream"
 	"github.com/tektoncd/pipeline/pkg/remote/oci"
 
-	"github.com/hacbs-contract/ec-cli/internal/acceptance/registry"
+	"github.com/hacbs-contract/ec-cli/acceptance/registry"
 )
 
 const version = "v1beta1"
