+++ codecov.yml
@@ -15,7 +15,7 @@
 # SPDX-License-Identifier: Apache-2.0
 ---
 ignore:
-  - "internal/acceptance"
+  - "acceptance"
 parsers:
   go:
     partials_as_hits: true
