+++ tools/tools.go
@@ -14,9 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build tools
-// +build tools
-
 // Tracks tooling dependencies, i.e. those that are not used by the code directly
 package tools
